docs: fill custom context page
