+++ docs/docs_src/getting_started/context/base_rabbit.py
@@ -1,7 +1,7 @@
 from faststream import Context, FastStream
 from faststream.rabbit import RabbitBroker
 
-broker = RabbitBroker()
+broker = RabbitBroker("amqp://guest:guest@localhost:5672/")
 app = FastStream(broker)
