+++ docs/docs_src/getting_started/context/custom_local_context_kafka.py
@@ -1,0 +1,24 @@
+from faststream import Context, FastStream, apply_types
+from faststream.kafka import KafkaBroker
+from faststream.kafka.annotations import ContextRepo, KafkaMessage
+
+broker = KafkaBroker("localhost:9092")
+app = FastStream(broker)
+
+
+@broker.subscriber("test-topic")
+async def handle(
+    msg: str,
+    message: KafkaMessage,
+    context: ContextRepo,
+):
+    with context.scope("correlation_id", message.correlation_id):
+        call()
+
+
+@apply_types
+def call(
+    message: KafkaMessage,
+    correlation_id=Context(),
+):
+    assert correlation_id == message.correlation_id
