+++ docs/includes/getting_started/context/custom_global.md
@@ -1,0 +1,9 @@
+=== "Kafka"
+    ```python linenums="1" hl_lines="9-10"
+    {!> docs_src/getting_started/context/custom_global_context_kafka.py [ln:1-5,16-18] !}
+    ```
+
+=== "RabbitMQ"
+    ```python linenums="1" hl_lines="9-10"
+    {!> docs_src/getting_started/context/custom_global_context_rabbit.py [ln:1-5,16-18] !}
+    ```
