+++ docs/includes/getting_started/context/custom_global_2.md
@@ -1,0 +1,9 @@
+=== "Kafka"
+    ```python linenums="1" hl_lines="4"
+    {!> docs_src/getting_started/context/custom_global_context_kafka.py [ln:8-13] !}
+    ```
+
+=== "RabbitMQ"
+    ```python linenums="1" hl_lines="4"
+    {!> docs_src/getting_started/context/custom_global_context_rabbit.py [ln:8-13] !}
+    ```
