+++ docs/includes/getting_started/context/custom_local.md
@@ -1,0 +1,9 @@
+=== "Kafka"
+    ```python linenums="1" hl_lines="15 19 21-22"
+    {!> docs_src/getting_started/context/custom_local_context_kafka.py !}
+    ```
+
+=== "RabbitMQ"
+    ```python linenums="1" hl_lines="15 19 21-22"
+    {!> docs_src/getting_started/context/custom_local_context_rabbit.py !}
+    ```
