+++ docs/includes/getting_started/context/manual_local.md
@@ -1,0 +1,9 @@
+=== "Kafka"
+    ```python linenums="1" hl_lines="1 14 24"
+    {!> docs_src/getting_started/context/manual_local_context_kafka.py !}
+    ```
+
+=== "RabbitMQ"
+    ```python linenums="1" hl_lines="1 14 24"
+    {!> docs_src/getting_started/context/manual_local_context_rabbit.py !}
+    ```
